Add attribute to select safety level Revamp call overhead test

+++ test/micro/org/openjdk/bench/jdk/incubator/foreign/libCallOverhead.c
@@ -34,8 +34,8 @@ EXPORT int identity(int x) {
 }
 
 typedef struct {
-    long long x;
-    long long y;
+    int x;
+    int y;
 } Point;
 
 EXPORT Point identity_struct(Point p) {
